+++ src/core/internal/index.ts
@@ -405,6 +405,7 @@ export class FrameElement {
       name: this.iFrameFormElement.fieldName,
       id: this.iFrameFormElement.iFrameName,
       placeholder: this.options.placeholder,
+      accept: this.options.accept,
       disabled: this.options.disabled ? true : undefined,
       readOnly: this.options.readOnly ? true : undefined,
       required: this.options.required ? true : undefined,