+++ tests/core-utils/reveal.test.js
@@ -85,13 +85,13 @@ describe('Reveal PureJs- get() Method Input', () => {
 
 describe("formatRecordsForClient fn test",()=>{
   test("only success records",()=>{
-    const testInput = {"records":[{"token":"7402-2242-2342-232","value":"231"}] }
+    const testInput = {"records":[{"token":"7402-2242-2342-232","value":"231", "valueType" : "STRING"}] }
     const fnResponse = formatRecordsForClient(testInput, {"7402-2242-2342-232": "231"});
     expect(fnResponse.success.length).toBe(1);
     expect(fnResponse.errors).toBeUndefined();
   });
   test("both success and error records",()=>{
-    const testInput = {"records":[{"token":"7402-2242-2342-232","value":"231"}],"errors":[{"token":"3232-6434-3253-4221"}]};
+    const testInput = {"records":[{"token":"7402-2242-2342-232","value":"231", "valueType" : "STRING"}],"errors":[{"token":"3232-6434-3253-4221"}]};
     const fnResponse = formatRecordsForClient(testInput,{"7402-2242-2342-232": "231"});
     expect(fnResponse.errors.length).toBe(1);
     expect(fnResponse.success.length).toBe(1);
@@ -111,7 +111,7 @@ describe("formatRecordsForIframe fn test",()=>{
     expect(fnResponse).toStrictEqual({});
   });
   test("with records should return token value object",()=>{
-    const testInput = {"records":[{token:"7823-323-242-2232",value:"token_value"}]};
+    const testInput = {"records":[{token:"7823-323-242-2232",value:"token_value","valueType" : "STRING"}]};
     const fnResponse = formatRecordsForIframe(testInput);
     expect(fnResponse).toStrictEqual({"7823-323-242-2232":"token_value"});
   });